Extract survey characteristics component Move Chooser into its own folder and extract the characteristics filter into its own component.
