+++ app/classifier/tasks/survey/chooser/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Chooser';
